QMOC

Documentation for QMOC.

Base.hash

QMOC.adoptedTrajectories

QMOC.allTrajectories

QMOC.archive

QMOC.archive

QMOC.archivedTrajectories

QMOC.backup

QMOC.boolArchived

QMOC.boolArchived

QMOC.boolComplete

QMOC.boolComplete

QMOC.cleanup

QMOC.countTrajectories

QMOC.loadMetadata

QMOC.loadMetadata

QMOC.missingTrajectories

QMOC.orphanedTrajectories

QMOC.parameter_circle

QMOC.parameter_distance

QMOC.parameter_full

QMOC.parameter_line

QMOC.parameter_wedge

QMOC.parametric_to_cartesian

QMOC.readTrajectory

QMOC.removeTrajectories

QMOC.standard_lines

QMOC.writeAverage

QMOC.writeMetadata

Base.hash — Method

hash(trajectory::Trajectory) -> UInt

Returns a hash of the trajectory properties.

QMOC.adoptedTrajectories — Method

adoptedTrajectories()

Returns a vector of all trajectory hashes in the filesystem that are in the metadata.

QMOC.allTrajectories — Method

allTrajectories()

Returns a vector of all trajectory hashes in the filesystem.

QMOC.archive — Method

archive(sim::Simulation) :: Nothing

Archives all trajectories in a simulation.

QMOC.archive — Method

archive(traj::Trajectory) :: Nothing

Archives a trajectory by first checking if it is complete and then archiving it.

QMOC.archivedTrajectories — Method

archivedTrajectories()

Returns a vector of all trajectory hashes in the archive.

QMOC.backup — Method

backup()

Backs up all metadata and measurements (including the archive) to a folder called "backup" in the data folder.

QMOC.boolArchived — Method

boolArchived(sim::Simulation) :: Bool

Checks if a simulation is archived, i.e. if all trajectories are archived.

QMOC.boolArchived — Method

boolArchived(traj::Trajectory) :: Bool

Checks if a trajectory is archived, i.e. if the measurement average is present in the archive.

QMOC.boolComplete — Method

boolComplete(sim::Simulation) :: Bool

Checks if a simulation is complete, i.e. if all linked trajectories are complete.

QMOC.boolComplete — Method

boolComplete(traj::Trajectory) :: Bool

Checks if a trajectory is complete, i.e. if all measurements are present. Automatically averages if all measurements are present but no average is found.

QMOC.cleanup — Method

cleanup()

Cleans up the filesystem by archiving all trajectories in the metadata and deleting all orphaned trajectories.

QMOC.countTrajectories — Method

countTrajectories(sim::Simulation) :: Int

Counts the number of trajectories in a simulation.

QMOC.loadMetadata — Method

loadMetadata(name::String)

Loads the metadata of a simulation from the filesystem and returns it. If the simulation name cannot be matched, it loads all metadata.

QMOC.loadMetadata — Method

loadMetadata()

Loads all metadata from the filesystem and returns a vector of simulations.

QMOC.missingTrajectories — Method

missingTrajectories(simulation::Simulation)

Returns a vector of all trajectories in a simulation that are not complete.

QMOC.orphanedTrajectories — Method

orphanedMeasurements()

Returns a vector of all trajectory hashes in the filesystem that are not in the metadata.

QMOC.parameter_circle — Method

parameter_circle(radius)

Returns the cartesian coordinates of a circle with radius `radius` around the isometric point.

QMOC.parameter_distance — Method

parameter_distance(param)

Returns the distance of a parameter set (px, py, pz) from the isotropic point in the parameter space.

QMOC.parameter_full — Method

parameter_full(resolution)

Returns a list of rational parameter sets (px, py, pz) filling the whole parameter space.

QMOC.parameter_line — Method

parameter_line(pointA, pointB, steps)

Returns a list of rational parameter sets (px, py, pz) on a line between pointA and pointB with steps number of points.

QMOC.parameter_wedge — Function

parameter_wedge(resolution, mode=:radial)

Returns a list of rational parameter sets (px, py, pz) filling the wedge between the isotropic point and the px=1 point, px=py point. Mode can be either :radial or :linear.

QMOC.parametric_to_cartesian — Method

parametric_to_cartesian(params)

Returns the cartesian coordinates of a parameter set (px, py, pz).

QMOC.readTrajectory — Method

readTrajectory(traj::Trajectory) :: Measurement

Reads a trajectory from the archive or from the filesystem. Assumes that the trajectory is complete.

QMOC.removeTrajectories — Method

removeTrajectories(sim::Simulation) :: Nothing

Removes all trajectories in a simulation from the filesystem.

QMOC.standard_lines — Method

standard_lines(resolution)

Returns a list of rational parameter sets (px, py, pz) on the two lines between the isotropic point and the px=1 point, px=py point.

QMOC.writeAverage — Method

writeAverage(traj::Trajectory) :: nothing

Overwrites a trajectory file containing a measurement series with its average.

QMOC.writeMetadata — Method

writeMetadata(sim::Simulation) :: Nothing

Writes the metadata of a simulation to the filesystem.